Improve handling of binary data.
Previously, when a textual format for a paste was requested, we
attempted to decode the data as UTF-8. On failure, we propagated the
error up which would result in a 500 error response. This is not ideal.

This commit improves on this by instead using lossy decoding which
decodes invalid UTF-8 into the replacement character. This results in a
valid UTF-8 string which can be displayed to the user.